# auth-service/controllers/swaggerDoc.yaml
openapi: "3.1.0"
info:
  title: "portfolio"
  version: "1.0.0"
  description: "portfolio app"
servers:
  - url: "__SWAGGER_API_BASE_URL__"
paths:
  /api/comments/{userId}:
    get:
      tags:
        - Comment Controller
      summary: Get paginated comments for userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of items per page.
      responses:
        '200':
          description: Comments retrieved successfully
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Comment does not have permission to access this resource
        '500':
          description: Internal server error - Something went wrong on the server side
    post:
      tags:
        - Comment Controller
      consumes: application/json
      summary: Create a new comment
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: string
              comment:
                type: string
      responses:
        '201':
          description: Comment created successfully
        '401':
          description: Unauthenticated
        '403':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/comments/{commentId}:
    delete:
      tags:
        - Comment Controller
      summary: Delete a comment by ID
      parameters:
        - in: path
          name: commentId
          required: true
          description: ID of the comment to delete
          schema:
            type: string
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          description: Unauthenticated
        '403':
          description: Unauthorized
        '404':
          description: Comment not found
        '500':
          description: Internal server error
  /api/users:
    post:
      tags:
        - User Controller
      consumes: application/json
      summary: Create a new user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: string
              userName:
                type: string
              email:
                type: string
              imageUrl:
                type: string
      responses:
        '200':
          description: User created successfully
        '500':
          description: Internal server error
  /api/users/{userId}:
    get:
      tags:
        - User Controller
      summary: Get user by userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: User found
        '404':
          description: User not found
        '500':
          description: Internal server error